Fix decodeURIComponent bug (#23)
* Fix decodeURIComponent bug

This fixes an issue with using `decodeURIComponent` on query string parameters that (may) have been truncated on the client side. If this truncation trims a percent encoding (ie, turning `%20` into `%2`, or `%E2%9A%A1` at any point), `decodeURIComponent` throws an error.

That would mean a client side bug report now turns into a server side bug report, and we've lost information. This is happening thousands of times a day.m

* Install strip-invalid-trailing-encoding

* Update strip-invalid-trailing-encoding

* Return decodeURIComponent's error stack when decoding fails